Added a new setting to control whether or not the uploaded CV's should be stored and available from backend.

+++ public/page_apply.php
@@ -42,7 +42,13 @@
 		$j = new Job($job_id);
 
 		$f = pathinfo($_FILES['apply_cv']['name']);
-		$filename = md5(time() . '_' . $_FILES['apply_cv']['name'] . uniqid()) . '.' . $f['extension'];
+		$basefilename = $f['filename'] . '_' . uniqid();
+		$suffix = 0;
+		do {
+			$suffix++;
+			$filename = $basefilename . '_' . $suffix . '.' . $f['extension'];
+		} while (file_exists(FILE_UPLOAD_DIR . $filename));
+
 		if (move_uploaded_file($_FILES['apply_cv']['tmp_name'], FILE_UPLOAD_DIR . $filename))
 		{
 			$attachment = $filename;
@@ -104,7 +110,7 @@
 		}
 		
 		// delete uploaded file (cleanup)
-		if ($attachment != '')
+		if (($attachment != '') && !($settings['keep_uploaded_cvs']))
 		{
 			unlink(APP_PATH . FILE_UPLOAD_DIR . $attachment);
 		}
